Fixed issue with pages loading side by side
Basic design for the login page now complete and clicking on the login button now takes you to the "portfolio overview" page.
