+++ app/app.css
@@ -97,7 +97,6 @@ body {
   
   .overview-container h1 {
     display: flex;
-    flex-direction: column;
     justify-content: center;
     align-items: center;
     margin-bottom: 20px;
